# server/openapi.yaml
openapi: "3.0.3"
info:
  title: ImmoChecker API
  version: "1.0.0"
  description: |
    API publique ImmoChecker pour créer et consulter des missions, EDL, lots et bâtiments.

    ## Authentification
    Toutes les requêtes nécessitent une clé API créée depuis **Paramètres → API & Intégrations**.

    ```
    Authorization: Bearer imk_live_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ## Permissions (scopes)
    Chaque clé API a un scope :
    - **`read`** — lecture seule (tous les `GET`)
    - **`read_write`** — lecture + écriture (`POST`, `PATCH`, `DELETE`)

    Une écriture avec une clé `read` renvoie un `403 FORBIDDEN_SCOPE`.

    ## Rate Limiting
    100 requêtes par minute par clé API. Les headers `RateLimit-*` indiquent les limites restantes.
    Dépassement → `429 RATE_LIMIT`.

    ## Pagination
    Toutes les listes sont paginées avec un curseur :
    ```
    GET /missions?limit=25&cursor=<id>
    ```
    La réponse contient `meta.cursor` (prochain curseur, `null` si fin de liste) et `meta.has_more`.
    `limit` par défaut = 25, max = 100.

    ## Webhooks
    Les webhooks (`mission.creee`, `mission.annulee`, `edl.signe`, …) ne se configurent pas via cette
    API publique mais depuis **Paramètres → API & Intégrations** (session admin). Pour la documentation
    du payload et de la signature HMAC, consulter le panneau Webhooks dans le back-office.

    ## ⚠️ Try it out
...
servers:
  - url: https://immo-check-production.up.railway.app/api/v1
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "imk_live_..."
      description: Clé API créée depuis Paramètres → API & Intégrations
  schemas:
    Meta:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
          description: Curseur pour la page suivante (`null` si fin de liste)
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        has_more:
          type: boolean
          example: true
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Message humain en français
          example: "Ressource introuvable"
        code:
          type: string
          description: Code machine stable (à utiliser en logique de branchement)
          example: "NOT_FOUND"
        details:
          description: Détails additionnels (ex. erreurs de validation)
          nullable: true
    Lot:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        designation:
          type: string
          example: "Appartement T3 — Lot 302"
        type_bien:
          type: string
          enum: [appartement, maison, local_commercial, bureau, parking, cave, autre]
          example: "appartement"
        etage:
          type: string
          nullable: true
          example: "3ème"
        surface:
          type: number
          nullable: true
          example: 67
        reference_interne:
          type: string
          nullable: true
          example: "BAIL-2025-0302"
        batiment:
          type: object
          properties:
            id:
              type: string
              format: uuid
              example: "b23f1a2c-9d8e-4b3a-a1c2-d3e4f5a6b7c8"
            designation:
              type: string
              example: "Résidence Victor Hugo"
            adresse:
              type: string
              example: "12 rue Victor Hugo, 75016 Paris"
...
    Mission:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "c3d4e5f6-a1b2-4c5d-8e9f-0a1b2c3d4e5f"
        reference:
          type: string
          example: "M-2026-0042"
        date_planifiee:
          type: string
          format: date
          example: "2026-04-18"
        heure_debut:
          type: string
          nullable: true
          example: "09:00"
        heure_fin:
          type: string
          nullable: true
          example: "11:30"
        statut:
          type: string
          enum: [planifiee, terminee, annulee]
          example: "planifiee"
        statut_rdv:
          type: string
          enum: [a_confirmer, confirme, reporte]
          example: "a_confirmer"
        avec_inventaire:
          type: boolean
          example: false
        commentaire:
          type: string
          nullable: true
          example: "RDV confirmé par le locataire"
        lot:
          $ref: '#/components/schemas/Lot'
        edls:
          type: array
          items:
            $ref: '#/components/schemas/EDL'
    CreateMissionInput:
      type: object
      required: [lot_id, sens, date_planifiee]
      properties:
        lot_id:
          type: string
          format: uuid
          description: UUID du lot sur lequel planifier la mission
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        sens:
          type: string
          enum: [entree, sortie, entree_sortie]
          description: "`entree_sortie` crée 2 EDL (un de chaque sens)"
          example: "entree"
        avec_inventaire:
          type: boolean
          default: false
          description: Ajouter un EDL d'inventaire en plus des EDL classiques
          example: false
        date_planifiee:
          type: string
          format: date
          example: "2026-04-18"
        heure_debut:
          type: string
          nullable: true
          example: "09:00"
        heure_fin:
          type: string
          nullable: true
          example: "11:00"
        commentaire:
          type: string
          nullable: true
          example: "Locataire prévenu, code interphone #1234B"
        technicien_id:
          type: string
          format: uuid
          nullable: true
          description: Assigner un technicien immédiatement (optionnel). Le statut reste `planifiee` — l'acceptation du technicien est tracée dans `mission_technicien.statut_invitation`.
          example: "5e8f9d3a-7b6c-4a1e-9f2d-3c4b5a6e7f8d"
        locataires:
          type: array
          description: Liste de `tiers_id` locataires (colocation ou bail simple)
          items:
            type: string
            format: uuid
          example: ["e5f6a7b8-c1d2-4e5f-0a1b-2c3d4e5f6a7b"]
        type_bail:
          type: string
          enum: [individuel, collectif]
          default: individuel
          description: "`individuel` = 1 EDL par locataire ; `collectif` = 1 EDL partagé par tous"
          example: "individuel"
    UpdateMissionInput:
      type: object
      description: |
        Tous les champs sont optionnels. Si la mission est `terminee`, seul `commentaire` est
        modifiable (les autres champs renvoient `403 MISSION_LOCKED`). Une mission `annulee`
        n'est plus modifiable du tout.
      properties:
        date_planifiee:
          type: string
          format: date
          example: "2026-04-19"
        heure_debut:
          type: string
          nullable: true
          example: "10:00"
        heure_fin:
          type: string
          nullable: true
          example: "12:00"
        commentaire:
          type: string
          nullable: true
          example: "Reporté à la demande du locataire"
        statut_rdv:
          type: string
          enum: [a_confirmer, confirme, reporte]
          example: "confirme"
    CancelMissionInput:
      type: object
      required: [motif]
      properties:
        motif:
          type: string
          description: Motif d'annulation obligatoire (audit). Impossible si un EDL est signé.
          example: "Demande du propriétaire"
    EDL:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d4e5f6a7-b1c2-4d5e-9f0a-1b2c3d4e5f6a"
        type:
          type: string
          enum: [etat_des_lieux, inventaire]
          example: "etat_des_lieux"
        sens:
          type: string
          enum: [entree, sortie]
          example: "entree"
        statut:
          type: string
          enum: [brouillon, signe, infructueux]
          example: "brouillon"
        mission_id:
          type: string
          format: uuid
          example: "c3d4e5f6-a1b2-4c5d-8e9f-0a1b2c3d4e5f"
        date_realisation:
          type: string
          format: date-time
          nullable: true
          example: "2026-04-18T09:30:00Z"
        date_signature:
          type: string
          format: date-time
          nullable: true
          example: "2026-04-18T11:00:00Z"
        motif_infructueux:
          type: string
          nullable: true
          description: Motif renseigné lorsque l'EDL est infructueux (absence locataire, refus, etc.)
          example: null
        pdf_url:
          type: string
          nullable: true
          description: URL pré-signée (24h) du PDF EDL standard. Non-null uniquement si `statut = signe`.
          example: "https://example.com/edl-abc123.pdf?exp=1743696000&sig=..."
        web_url:
          type: string
          nullable: true
          description: URL persistante de consultation web. Non-null uniquement si `statut = signe`.
          example: "https://example.com/edl-abc123"
        pdf_url_legal:
          type: string
          nullable: true
          description: URL pré-signée (24h) du PDF version légale. Non-null uniquement si `statut = signe`.
          example: null
        web_url_legal:
          type: string
          nullable: true
          description: URL persistante de consultation légale. Non-null uniquement si `statut = signe`.
          example: null
        url_verification:
          type: string
          nullable: true
          description: Page publique de vérification (QR code, scellé électronique). Non-null uniquement si `statut = signe`.
          example: null
        locataires:
          type: array
          nullable: true
          items:
            type: object
            properties:
              tiers_id:
                type: string
                format: uuid
                example: "e5f6a7b8-c1d2-4e5f-0a1b-2c3d4e5f6a7b"
              nom:
                type: string
                example: "Leroy"
              prenom:
                type: string
                nullable: true
                example: "Pierre"
              role_locataire:
                type: string
                enum: [entrant, sortant]
                example: "entrant"
        cles:
          type: array
          nullable: true
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
                example: "f6a7b8c9-d1e2-4f5a-1b2c-3d4e5f6a7b8c"
              type_cle:
                type: string
                example: "cle_principale"
              quantite:
                type: integer
                example: 2
              statut:
                type: string
                enum: [remise, a_deposer, deposee]
                example: "remise"
              lieu_depot:
                type: string
                nullable: true
                example: "Agence principale"
    PaginatedMissions:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Mission' }
        meta: { $ref: '#/components/schemas/Meta' }
...
  responses:
    Unauthorized:
      description: Clé API manquante ou invalide
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missing:
              summary: En-tête Authorization absent
              value: { error: "Clé API requise", code: "UNAUTHORIZED" }
            invalid:
              summary: Clé inconnue ou révoquée
              value: { error: "Clé API invalide", code: "INVALID_API_KEY" }
    Forbidden:
      description: Clé valide mais scope insuffisant (ex. `read` tentant un POST)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            readOnly:
              value: { error: "Clé API en lecture seule", code: "FORBIDDEN_SCOPE" }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            default:
              value: { error: "Mission introuvable", code: "NOT_FOUND" }
    RateLimit:
      description: Trop de requêtes (100/min par clé)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            default:
              value: { error: "Trop de requêtes", code: "RATE_LIMIT" }
    Validation422:
      description: Données invalides (champ manquant, format incorrect, valeur enum inconnue)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: array
                          items: { type: string }
                        message:
                          type: string
          examples:
            default:
              value:
                error: "Validation échouée"
                code: "VALIDATION_ERROR"
                details:
                  - path: ["date_planifiee"]
                    message: "Invalid date"
paths:
  /missions:
    get:
      summary: Lister les missions
      operationId: listMissions
      tags: [Missions]
      parameters:
        - in: query
          name: statut
          schema:
            type: string
            enum: [planifiee, terminee, annulee]
        - in: query
          name: date_from
          schema: { type: string, format: date }
        - in: query
          name: date_to
          schema: { type: string, format: date }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 25, maximum: 100 }
      responses:
        '200':
          description: Liste paginée de missions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedMissions' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimit' }

    post:
      summary: Créer une mission
      operationId: createMission
      tags: [Missions]
      description: |
        Crée une mission et ses EDL associés (1 par sens pour un bail simple, 1 par locataire
        pour un bail `individuel`, 1 partagé pour un bail `collectif`).

        **Exemple curl :**
        ```bash
        curl -X POST https://immo-check-production.up.railway.app/api/v1/missions \
          -H "Authorization: Bearer imk_live_..." \
          -H "Content-Type: application/json" \
          -d '{"lot_id":"...","sens":"entree","date_planifiee":"2026-04-18"}'
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMissionInput' }
      responses:
        '201':
          description: Mission créée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mission' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Validation422' }
        '429': { $ref: '#/components/responses/RateLimit' }
  /missions/{id}:
    get:
      summary: Détail d'une mission
      operationId: getMission
      tags: [Missions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Détail mission avec EDLs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mission' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

    patch:
      summary: Modifier une mission
      operationId: patchMission
      tags: [Missions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMissionInput' }
      responses:
        '200':
          description: Mission mise à jour
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mission' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: "Mission verrouillée (`terminee` → seul `commentaire` modifiable ; `annulee` → interdit)"
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                locked:
                  value: { error: "Mission terminée : seul le commentaire est modifiable", code: "MISSION_LOCKED" }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Validation422' }

    delete:
      summary: Annuler une mission
      operationId: cancelMission
      tags: [Missions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CancelMissionInput' }
      responses:
        '200':
          description: Mission annulée
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled: { type: boolean, example: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Impossible d'annuler une mission avec des EDL signés
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              examples:
                signedEdl:
                  value: { error: "Impossible d'annuler : au moins un EDL est signé", code: "MISSION_LOCKED" }
        '422': { $ref: '#/components/responses/Validation422' }
  /missions/{id}/edl-inventaires:
    get:
      summary: EDL d'une mission
      operationId: getMissionEdls
      tags: [Missions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Liste des EDL de la mission
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EDL' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...